set namespaces for Variables
