+++ src/sidebar.ts
@@ -3,13 +3,13 @@
 
 import { SplitPanel } from '@phosphor/widgets';
 
-import { BreakPointsWidget } from './breakpoints';
+import { BreakPoints } from './breakpoints';
 
 import { Debugger } from './debugger';
 
-import { CallstackWidget } from './callstack';
+import { Callstack } from './callstack';
 
-import { VariablesWidget } from './variables';
+import { Variables } from './variables';
 
 export class DebuggerSidebar extends SplitPanel {
   constructor(model: Debugger.Model | null) {
@@ -18,20 +18,20 @@ export class DebuggerSidebar extends SplitPanel {
     this.orientation = 'vertical';
     this.addClass('jp-DebuggerSidebar');
 
-    this.variables = new VariablesWidget();
-    this.callstack = new CallstackWidget();
-    this.breakPoints = new BreakPointsWidget();
+    this.variables = new Variables();
+    this.callstack = new Callstack();
+    this.breakPoints = new BreakPoints();
 
     this.addWidget(this.variables);
     this.addWidget(this.callstack);
     this.addWidget(this.breakPoints);
   }
 
-  readonly variables: VariablesWidget;
+  readonly variables: Variables;
 
-  readonly callstack: CallstackWidget;
+  readonly callstack: Callstack;
 
-  readonly breakPoints: BreakPointsWidget;
+  readonly breakPoints: BreakPoints;
 
   get model(): Debugger.Model | null {
     return this._model;
