+++ src/variables/description/index.ts
