+++ src/variables/toggle/useOutsideClick.tsx
